Move the override to lab subclass to limit blast radius
