+++ packages/application/src/frontend.ts
@@ -203,72 +203,6 @@ export abstract class JupyterFrontEnd<
     }
   }
 
-  /**
-   * Override keydown handling to prevent command shortcuts from preventing user input.
-   *
-   * This introduces a slight delay to the command invocation, but no delay to user input.
-   */
-  protected evtKeydown(event: KeyboardEvent): void {
-    // Process select keys which may call `preventDefault()` immediately
-    if (
-      ['Tab', 'ArrowDown', 'ArrowUp', 'ArrowRight', 'ArrowLeft'].includes(
-        event.key
-      )
-    ) {
-      return this.commands.processKeydownEvent(event);
-    }
-    // Process remaining events conditionally, depending on whether they would lead to text insertion
-    const causesInputPromise = Promise.race([
-      new Promise(resolve => {
-        if (!event.target) {
-          return resolve(false);
-        }
-        event.target.addEventListener(
-          'beforeinput',
-          (inputEvent: InputEvent) => {
-            switch (inputEvent.inputType) {
-              case 'historyUndo':
-              case 'historyRedo': {
-                if (
-                  inputEvent.target instanceof Element &&
-                  inputEvent.target.closest('[data-jp-undoer]')
-                ) {
-                  // Allow to use custom undo/redo bindings on `jpUndoer`s
-                  inputEvent.preventDefault();
-                  return resolve(false);
-                }
-                break;
-              }
-              case 'insertLineBreak': {
-                if (
-                  inputEvent.target instanceof Element &&
-                  inputEvent.target.closest('.jp-Cell')
-                ) {
-                  // Allow to override the default action of Shift + Enter on cells as this is used for cell execution
-                  inputEvent.preventDefault();
-                  return resolve(false);
-                }
-                break;
-              }
-            }
-            return resolve(true);
-          },
-          { once: true }
-        );
-      }),
-      new Promise(resolve => {
-        setTimeout(() => resolve(false), Private.INPUT_GUARD_TIMEOUT);
-      })
-    ]);
-    causesInputPromise
-      .then(willCauseInput => {
-        if (!willCauseInput) {
-          this.commands.processKeydownEvent(event);
-        }
-      })
-      .catch(console.warn);
-  }
-
   private _contextMenuEvent: MouseEvent;
   private _format: U;
   private _formatChanged = new Signal<this, U>(this);
@@ -471,16 +405,6 @@ namespace Private {
   export function suppressContextMenu(element: HTMLElement): boolean {
     return element.closest('[data-jp-suppress-context-menu]') !== null;
   }
-
-  /**
-   * The delay for invoking a command introduced by user input guard.
-   * Decreasing this value may lead to commands incorrectly triggering
-   * on user input. Increasing this value will lead to longer delay for
-   * command invocation. Note that user input is never delayed.
-   *
-   * The value represents the number in milliseconds.
-   */
-  export const INPUT_GUARD_TIMEOUT = 10;
 }
 
 /**
